frameworks/av: Fix playback for LPA audio when using gapless api
LPAPlayer in its current state shows a very odd behavior.
It plays one song over DSP processor and the next is played with software decoding.
This is because there are two open instances of awesomeplayer.
At the moment awesomeplayer checks if DSP processor is free, if it's not, DSP falls back to software.
This patch fixes that by delaying that check to the actual call of play(),
so the awesomeplayer which currently holds the DSP processor can free it.
This is also done earlier than before.

Based on the cm-10.2 tunnel audio commit by Hendrik Borghorst <hendrikborghorst@gmail.com>:
http://review.cyanogenmod.org/46607

Change-Id: Ib342a154dd47842fd085082e65d24fe7e322ff5b